WIP: Plugin system for execution engines.
