+++ cosmic_ray/execution/execution_engine.py
@@ -1,0 +1,7 @@
+import abc
+
+
+class ExecutionEngine(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def __call__(self, timeout, pending_work, config):
+        raise NotImplemented()
